+++ src/TurtlePath/Models/Requests/BaseRequest.cs
@@ -1,0 +1,17 @@
+namespace TurtlePath.Models.Requests
+{
+    using TurtlePath.Domain.Identifier;
+
+    /// <summary>
+    /// Represents the base request model containing a unique identifier.
+    /// </summary>
+    public abstract class BaseRequest
+    {
+        /// <summary>
+        /// Gets or sets the unique identifier for the resource.
+        /// </summary>
+        public CId Id { get; set; }
+    }
+}
+
+